pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fragmind"
version = "0.1.0"
description = "于 AI Agent 的碎片化思维整理与日记生成系统"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Joyu", email = "joyu.gbc@outlook.com" }
]
dependencies = [
    "PyQt6>=6.6.1",
    "python-dotenv>=1.0.0",
    "pydantic-ai>=0.0.14",
    "logfire>=0.46.0",
    "openai>=1.10.0",
    "devtools>=0.12.0",
    "nest_asyncio>=1.6.0",
    "qasync>=0.27.1",
    "httpx[socks]>=0.28.1",
]


[tool.hatch.build.targets.wheel]
packages = ["src"]
